Feito um upload para o logo da equipe, consequentemente mudando o controller Equipes e adicionando um input de arquivos no FormEquipes

+++ application/controllers/Equipes.php
@@ -1,11 +1,12 @@
 <?php
 
 
-class Equipes extends CI_Controller{
+class Equipes extends CI_Controller
+{
 
-    public function index(){
+    public function index()
+    {
         $this->listar();
-        
     }
 
     public function __construct()
@@ -32,11 +33,26 @@ class Equipes extends CI_Controller{
             $this->load->view('FormEquipes');
             $this->load->view('Footer');
         } else {
-            
+
 
             $data = array(
                 'nome' => $this->input->post('nome')
             );
+
+            $config['upload_path']          = './uploads/';
+            $config['allowed_types']        = 'gif|jpg|png';
+            $config['max_width']            = 1024;
+            $config['max_height']           = 768;
+            $config['encrypt_name']         = true;
+            $this->load->library('upload', $config);
+            if (!$this->upload->do_upload('userfile')) {
+                $error = ($this->upload->display_errors());
+                $this->session->set_flashdata('mensagem', "$error");
+                redirect('Equipes/Listar');
+                exit;
+            } else {
+                $data['imagem'] = $this->upload->data('file_name');
+            }
 
             if ($this->Equipes_model->insert($data)) {
                 $this->session->set_flashdata('mensagem', 'Equipe cadastrada com sucesso!!!');
@@ -51,7 +67,7 @@ class Equipes extends CI_Controller{
     public function alterar($id)
     {
         if ($id > 0) {
-        
+
             $this->form_validation->set_rules('nome', 'nome', 'required');
 
             if ($this->form_validation->run() == FALSE) {
@@ -63,6 +79,20 @@ class Equipes extends CI_Controller{
                 $data = array(
                     'nome' => $this->input->post('nome')
                 );
+                $config['upload_path']          = './application/uploads/';
+                $config['allowed_types']        = 'gif|jpg|png';
+                $config['max_width']            = 1024;
+                $config['max_height']           = 768;
+                $config['encrypt_name']         = true;
+                $this->load->library('upload', $config);
+                if (!$this->upload->do_upload('userfile')) {
+                    $error = ($this->upload->display_errors());
+                    $this->session->set_flashdata('mensagem', "$error");
+                    redirect('Equipes/Listar');
+                    exit;
+                } else {
+                    $data['imagem'] = $this->upload->data('file_name');
+                }
                 if ($this->Equipes_model->update($id, $data)) {
                     $this->session->set_flashdata('mensagem', 'Equipe alterada com sucesso!!!');
                     redirect('Equipes/Listar');
@@ -79,11 +109,14 @@ class Equipes extends CI_Controller{
     public function deletar($id)
     {
         if ($id > 0) {
-            
-            if($this->Equipes_model->delete($id)){
-                $this->session->set_flashdata('mensagem', 'Equipe deletada com sucesso!!!');
-            }else{
-                $this->session->set_flashdata('mensagem', 'Erro ao deletar equipe...');
+            $equipe = $this->Equipes_model->getOne($id);
+            if ($equipe) {
+                unlink('./uploads/' . $equipe->imagem);
+                if ($this->Equipes_model->delete($id)) {
+                    $this->session->set_flashdata('mensagem', 'Equipe deletada com sucesso!!!');
+                } else {
+                    $this->session->set_flashdata('mensagem', 'Erro ao deletar equipe...');
+                }
             }
         }
         redirect('Equipes/listar');
